refactor(docs): extract chart page content for SSR metadata
- Refactor all chart component pages to server components
- Extract interactive content to separate client components
- Add SEO metadata with keywords and canonical URLs
- Add breadcrumb schemas for search engine navigation
